Fix requested changes
 This commit fixes errors and adds validate method for inputs
 - Throws an error incase a user does not exist
 - Add a validator function to validate email and password
 - Throws an error incase an error occurs
   when fetching a user from the database
