+++ server/src/graphql/resolvers/mutation.js
@@ -1,33 +1,27 @@
-import {UserInputError} from 'apollo-server';
 import bcrypt from "bcrypt";
-import isEmail from "validator/lib/isEmail";
-import isLength from "validator/lib/isLength"
-import helpers from '../../helpers'
+import { generateToken, validateInput } from '../../helpers'
 
 export const Mutation = {
   register: async (_, { email, password }, { client }) => {
-    if(!isEmail(email)){
-      throw new UserInputError('Incorrect email format! 😢');
-    }
-    if(!isLength(password, 8)){
-      throw new UserInputError('Password cannot be left blank or less than 8 characters! 😢');
-    }
+    validateInput(email, password);
     const salt = await bcrypt.genSalt();
     const hashedPassword = await bcrypt.hash(password, salt);
     return await client.createUser({ email, password: hashedPassword });
   },
   login: async (_, { email, password }, { client })=>{
+    validateInput(email, password);
     const user = await client.getUserByEmail(email);
+    if(!user ) throw Error("User not found");
+    if(typeof user === "string") throw Error(user);
     const match = await bcrypt.compare(password, user.password);
-    if(match){
-      return {
-        token :() => helpers.generateToken(user.email),
+    if(!match)throw new Error("Incorrect password");
+    const token = generateToken(user.email);
+    return {
+        token: token,
         user:{
           id: user.user_id,
           email: user.email
         }
       }
     }
-    throw new Error("Incorrect password");
-  }
 };
